+++ db/flush_job.cc
@@ -182,7 +182,8 @@ Status FlushJob::WriteLevel0Table(const autovector<MemTable*>& mems,
 
       s = BuildTable(dbname_, db_options_.env, *cfd_->ioptions(), env_options_,
                      cfd_->table_cache(), iter.get(), &meta,
-                     cfd_->internal_comparator(), newest_snapshot_,
+                     cfd_->internal_comparator(),
+                     cfd_->int_tbl_prop_collector_factories(), newest_snapshot_,
                      earliest_seqno_in_memtable, output_compression_,
                      cfd_->ioptions()->compression_opts, Env::IO_HIGH);
       LogFlush(db_options_.info_log);