+++ table/adaptive_table_factory.h
@@ -39,12 +39,9 @@ class AdaptiveTableFactory : public TableFactory {
       unique_ptr<RandomAccessFile>&& file, uint64_t file_size,
       unique_ptr<TableReader>* table) const override;
 
-  TableBuilder* NewTableBuilder(const ImmutableCFOptions& ioptions,
-                                const InternalKeyComparator& icomparator,
-                                WritableFile* file,
-                                const CompressionType compression_type,
-                                const CompressionOptions& compression_opts,
-                                const bool skip_filters) const override;
+  TableBuilder* NewTableBuilder(
+      const TableBuilderOptions& table_builder_options,
+      WritableFile* file) const override;
 
   // Sanitizes the specified DB Options.
   Status SanitizeOptions(const DBOptions& db_opts,