+++ table/cuckoo_table_factory.h
@@ -59,11 +59,9 @@ class CuckooTableFactory : public TableFactory {
       unique_ptr<RandomAccessFile>&& file, uint64_t file_size,
       unique_ptr<TableReader>* table) const override;
 
-  TableBuilder* NewTableBuilder(const ImmutableCFOptions& options,
-                                const InternalKeyComparator& icomparator,
-                                WritableFile* file, const CompressionType,
-                                const CompressionOptions&,
-                                const bool skip_filters = false) const override;
+  TableBuilder* NewTableBuilder(
+      const TableBuilderOptions& table_builder_options,
+      WritableFile* file) const override;
 
   // Sanitizes the specified DB Options.
   Status SanitizeOptions(const DBOptions& db_opts,