+++ table/mock_table.h
@@ -149,12 +149,9 @@ class MockTableFactory : public TableFactory {
                                const InternalKeyComparator& internal_key,
                                unique_ptr<RandomAccessFile>&& file, uint64_t file_size,
                                unique_ptr<TableReader>* table_reader) const override;
-  TableBuilder* NewTableBuilder(const ImmutableCFOptions& ioptions,
-                                const InternalKeyComparator& internal_key,
-                                WritableFile* file,
-                                const CompressionType compression_type,
-                                const CompressionOptions& compression_opts,
-                                const bool skip_filters = false) const override;
+  TableBuilder* NewTableBuilder(
+      const TableBuilderOptions& table_builder_options,
+      WritableFile* file) const override;
 
   // This function will directly create mock table instead of going through
   // MockTableBuilder. MockFileContents has to have a format of <internal_key,